backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
cryptography==46.0.5
Deprecated==1.3.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
google-auth==2.49.0
gunicorn==21.2.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
limits==4.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==1.26.2
ordered-set==4.1.0
packaging==24.2
pandas==2.1.4
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2026.1.post1
requests==2.32.5
rich==13.9.4
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.13.1
six==1.17.0
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.0.1
wrapt==2.1.2
zipp==3.23.0
